% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDE_sim_glm.R
\name{SDE_glm4}
\alias{SDE_glm4}
\title{SDE_tmle3}
\usage{
SDE_glm4(data, covariates, truth = NULL, truncate = list(lower = 1e-04,
  upper = 0.9999), B = 500)
}
\arguments{
\item{data, }{data.frame of variables in time ordering from left to right}

\item{covariates, }{list of covariates for each necessary model, going backwards from the
outcome, Y, M, Z, A, W, S where S is the binary site, W are confounders, A is the treatment
Z is the intermediary confounder (binary) and M is the mediator, Y is the outcome.}

\item{truth}{for testing purposes input a list with names f_W, f_S, f_Z and f_Y models representing
and the corresponding elements are functions of appropriate variables so as to be able to
generate the truth and check the estimator's performance.  Default is NULL}

\item{truncate, }{a list with elements lower and upper to truncate the various p-scores
not functional at present}

\item{a, }{the treatment intervention of interest}

\item{a_star, }{the treatment intervention of the stochastic model for Mediator, M}

\item{sl}{the sl3 superlearner defined, see sl3 documentation for defining a superlearner
and the example below}

\item{V}{number of folds for cross-validation (fixed to 10 for now)}
}
\value{
a list with a CI for the estimate, and estimate using linear main terms MLE
gcomp formula (est_mle), the influence curve (IC), the superlearner coefficients for
the Y model and the QZ model (SL_coef)
}
\description{
computes the sequential regression, targeted maximum likelihood estimate
for the stochastic direct effect or stochastic indirect effect when the outcome and
mediator model are only available on site 1 (S = 1).  This is a data adaptive parameter
as the stochastic direct effect has a model for the mediator is determined on the data for
site 1.
}
